Revert "move FindVespaUser(), add FindVespaUidAndGid()"
