+++ client/go/vespa/load_env.go
@@ -8,6 +8,7 @@ import (
 	"bufio"
 	"fmt"
 	"os"
+	"os/user"
 	"strings"
 )
 
@@ -20,10 +21,40 @@ func LoadDefaultEnv() error {
 func ExportDefaultEnvToSh() error {
 	holder := newShellEnvExporter()
 	err := loadDefaultEnvTo(holder)
-	holder.overrideVar("VESPA_HOME", FindHome())
-	holder.overrideVar("VESPA_USER", FindVespaUser())
 	holder.dump()
 	return err
+}
+
+// Which user should vespa services run as?  If current user is root,
+// we want to change to some non-privileged user.
+// Should be run after LoadDefaultEnv() which possibly loads VESPA_USER
+func FindVespaUser() string {
+	uName := os.Getenv("VESPA_USER")
+	if uName != "" {
+		// no check here, assume valid
+		return uName
+	}
+	if os.Getuid() == 0 {
+		u, err := user.Lookup("vespa")
+		if err == nil {
+			uName = u.Username
+		} else {
+			u, err = user.Lookup("nobody")
+			if err == nil {
+				uName = u.Username
+			}
+		}
+	}
+	if uName == "" {
+		u, err := user.Current()
+		if err == nil {
+			uName = u.Username
+		}
+	}
+	if uName != "" {
+		os.Setenv("VESPA_USER", uName)
+	}
+	return uName
 }
 
 type loadEnvReceiver interface {
